pyproject.toml
[project]
name = "vps-config"
version = "0.1.0"
description = "Personal VPS configuration with Ansible"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "ansible>=6.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.0.0",
    "requests>=2.28.0",
]

[project.scripts]
test-local = "scripts.test_local:main"
validate = "scripts.validate:main"
deploy = "scripts.deployment.deploy:main"
health-check = "scripts.utilities.health_check:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["scripts"]